+++ src/routing.rs
@@ -1,5 +1,6 @@
 //! Routing between [`Service`]s.
 
+use self::future::{BoxRouteFuture, EmptyRouterFuture, RouteFuture};
 use crate::{
     body::{box_body, BoxBody},
     buffer::MpscBuffer,
@@ -26,11 +27,6 @@ use tower::{
 use tower_http::map_response_body::MapResponseBodyLayer;
 
 pub mod future;
-
-// for backwards compatibility
-// TODO: remove these in 0.2
-#[doc(hidden)]
-pub use self::future::{BoxRouteFuture, EmptyRouterFuture, RouteFuture};
 
 /// A filter that matches one or more HTTP methods.
 #[derive(Debug, Copy, Clone)]